requirements.txt
# Spectr PDF — Windows Desktop App Requirements
# pip install -r requirements.txt

# ── GUI ───────────────────────────────────────────────────────────────────────
PyQt6>=6.5.0

# ── PDF engine ────────────────────────────────────────────────────────────────
pymupdf>=1.23.0
pikepdf>=8.0.0

# ── Image processing ──────────────────────────────────────────────────────────
pillow>=10.0.0
opencv-python-headless>=4.8.0

# ── Digital signatures ────────────────────────────────────────────────────────
pyhanko>=0.20.0
cryptography>=41.0.0

# ── CAC / smart card ──────────────────────────────────────────────────────────
PyKCS11>=1.5.0

# ── OCR ───────────────────────────────────────────────────────────────────────
pytesseract>=0.3.10

# ── Document conversion ───────────────────────────────────────────────────────
python-docx>=1.0.0

# ── Build ─────────────────────────────────────────────────────────────────────
# pyinstaller>=6.0.0   # install separately: pip install pyinstaller

# ── pyhanko optional deps (eliminate build warnings) ──────────────────────────
xsdata>=23.0.0
aiohttp>=3.8.0
